Admin-Dashboard - add product / edit product

+++ api/admin/product/delete.php
@@ -24,6 +24,19 @@ if(!$id)
 }
 
 
+$stmt = $conn->prepare("SELECT id FROM order_items WHERE product_id = ? LIMIT 1");
+$stmt->bind_param("i", $id);
+$stmt->execute();
+$stmt->store_result();
+
+if ($stmt->num_rows > 0) {
+    http_response_code(400); // Bad Request or 409 Conflict
+    echo json_encode(['status' => 'error', 'message' => 'Product currently not deletable, it is ordered by user']);
+    exit;
+}
+$stmt->close();
+
+
 $stmt = $conn->prepare("DELETE FROM products WHERE id = ?");
 $stmt->bind_param("i", $id);
